Game post-FINAL version
If additional info/feedback will be received, further changes will be made.

The game does not have an increasing difficulty, but the survival the limited amount of lives counts as enough.
As the player progresses throughout the levels, the lives are not getting replenished, therefore the loss comes by making mistakes.