Character ANIMATIONS and WEAPON spawn/kill/switch
The character is following the direction of the keys and has moving, idle and attack animations and all the right orders.

Added weapons which will spawn the sprite for when the player is updated in the attack status and the weapons gets destroyed from the position when the animation is finished.
5 weapons have intended in a list and the another input key (q) has been added for switching inside the list, so the player can switch between weapons.

There is a cooldown for both attack and switching weapons.

Follow up:
Build UI interface (healthbar/manabar and so on.)
Magic
Enemies